ohos: Extras: return native ids as qint64, not double
tryGetNativeWindowId() and tryGetScreenDisplayId() returned
std::optional<double>. The values are read through the ETS/JS API,
where every number is a double, which is how they surfaced as double
in the external API - but the ids are integral, so double is a
surprising and lossy-looking type for them.

The native SDK confirms the ids are integral, but with no single
canonical type: window id is int32_t in the window manager
(oh_window.h) yet uint32_t/uint64_t in the PiP and vsync APIs,
display id is int32_t in start_options, int64_t in
OH_WindowManager_GetAllWindowLayoutInfoList(), and uint32_t/uint64_t
in the display manager. Because the native types are mixed in both
width and signedness, use qint64: it is signed (matching Qt
convention and QAndroidScreen::displayId()) and wide enough to hold
every native form without narrowing. Use the existing
convertDoubleTo() helper to validate the value first and warn
instead of returning a invalid id.

Convert at the boundary, private interfaces still deal in double.

Task-number: QTBUG-147194
Pick-to: 6.12

+++ src/plugins/platforms/ohos/extras/qohoswindowutils_p.h
@@ -15,11 +15,12 @@
 // We mean it.
 //
 
+#include <QtHarmonyExtras/private/qtharmonyextrasglobal_p.h>
 #include <QtGui/qcolor.h>
 #include <QtGui/qscreen.h>
 #include <QtGui/qwindow.h>
-#include <QtHarmonyExtras/private/qtharmonyextrasglobal_p.h>
 #include <QtWidgets/qwidget.h>
+#include <QtCore/qtypes.h>
 #include <optional>
 
 QT_BEGIN_NAMESPACE
@@ -59,9 +60,9 @@ Q_HARMONYEXTRAS_EXPORT void setMainWindowGeometryPersistenceHint(WindowGeometryP
 
 }
 
-Q_HARMONYEXTRAS_EXPORT std::optional<double> tryGetNativeWindowId(QWindow *window);
+Q_HARMONYEXTRAS_EXPORT std::optional<qint64> tryGetNativeWindowId(QWindow *window);
 
-Q_HARMONYEXTRAS_EXPORT std::optional<double> tryGetScreenDisplayId(QScreen *screen);
+Q_HARMONYEXTRAS_EXPORT std::optional<qint64> tryGetScreenDisplayId(QScreen *screen);
 
 }
